+++ packages/gakumas-engine/engine/TurnManager.js
@@ -118,7 +118,13 @@ export default class TurnManager extends EngineComponent {
     this.engine.effectManager.triggerEffectsForPhase(state, "endOfTurn");
 
     // Good impression
-    this.engine.effectManager.triggerEffects(state, GOOD_IMPRESSION_EFFECTS);
+    const goodImpressionTurnsTimes = state[S.goodImpressionTurnsTimesBuffs].reduce(
+      (acc, cur) => acc + cur.amount,
+      1
+    );
+    for (let i = 0; i < goodImpressionTurnsTimes; i++) {
+      this.engine.effectManager.triggerEffects(state, GOOD_IMPRESSION_EFFECTS);
+    }
 
     // Decrement buff turns
     this.engine.buffManager.decrementBuffTurns(state);