+++ packages/gakumas-engine/strategies/HeuristicStrategy.js
@@ -122,7 +122,7 @@ export default class HeuristicStrategy extends BaseStrategy {
       (acc, cur) =>
         acc +
         (config.typeMultipliers[cur] * config.stage.turnCounts[cur]) /
-          config.stage.turnCount,
+        config.stage.turnCount,
       0
     );
   }
@@ -281,6 +281,15 @@ export default class HeuristicStrategy extends BaseStrategy {
     // Good impression turns effects buffs
     score +=
       state[S.goodImpressionTurnsEffectBuffs].reduce(
+        (acc, cur) => acc + cur.amount * (cur.turns || state[S.turnsRemaining]),
+        0
+      ) *
+      state[S.goodImpressionTurns] *
+      this.goodImpressionTurnsMultiplier;
+
+    // Good impression turns times buffs
+    score +=
+      state[S.goodImpressionTurnsTimesBuffs].reduce(
         (acc, cur) => acc + cur.amount * (cur.turns || state[S.turnsRemaining]),
         0
       ) *