+++ gakumas-tools/simulator/CHANGELOG.md
@@ -9,6 +9,12 @@ Dates are based on Eastern Time Zone.
 
 
 
+
+## 2026-03-20
+
+- Add support for good impression times effect"
+- Add support for good impression times effect
+
 ## 2026-03-18
 
 - Add new stages, cards, p-items. Change handling of card counting effects